# Avoid decompressing just to calculate external size