Use snappy's `uncompressed_length` and external binary format's binary spec to
get uncompressed size.

http://erlang.org/doc/apps/erts/erl_ext_dist.html