@eiri 2017-09-22
Looks good to me, but changing `?term_size/1` scares me. @davisp - second opinion, please?