@eiri 2017-09-22
Also addresses #829, for sake of references.