@janl 2017-09-22
I’ve independently and accidentally developed nearly the same patch. I avoided redefining `?term_size`, but this lgtm if @davisp agrees.

Would be nice to get a few more numbers on how much faster this is cc @chewbranca. This could play a nice role in the 2.2 announcement.