@janl 2017-09-22
Ran more numbers, a 1GB database (npm registry subset) with the whole database file in the fs cache (e.g. no read activity on the disk), seeing improvements between 10% and 30% in both CPU and wall clock time (normalised over 3 runs each), and corresponding higher write throughput to disk. All on my laptop so highly unscientific, but enough of a difference to mark as significant.