@nickva 2017-09-22
Good call on benchmarking this.

I have a compaction benchmark script and with `deflate_9` compression and large docs it shows good improvement about 40% or so.

https://gist.github.com/nickva/e87aa2f7f896805bfee36a044a3800b0

So it makes 100k docs of size 10k each. Then times compaction.

*master* 

```
$ ./compact_bench.py -n 100000 -s 10000

**************** num=100000,size=10000 ****************
 Updating   : 1059.3s docs/s:94 revs/s:94 fsize:850620615
 Compacting : 35.5s docs/s:2816 revs/s:2816 fsize:773591239
```

*PR branch*

```
$ ./compact_bench.py -n 100000 -s 10000

**************** num=100000,size=10000 ****************
 Updating   : 1000.7s docs/s:99 revs/s:99 fsize:850620615
 Compacting : 19.6s docs/s:5091 revs/s:5091 fsize:773599431
````