@davisp 2017-09-25
@eiri Actually not having minor_version in there is a bit of an oversight as we use minor_version everywhere when doing term_to_binary calls before writing to disk (as we should).

+1

Great find!